+++ Ping-Application/src/main/kotlin/com/ping/application/member/dto/TokenResponse.kt
@@ -1,0 +1,5 @@
+package com.ping.application.member.dto
+
+data class TokenResponse(
+    val accessToken: String,
+)
